Affichage espace restant dans mon espace Découpage du nom de fichier quand trop long Affichage erreur quand user/mdp incorrect Affichage message quand user logout Création / paiement / transaction fonctionnel Correction bug lors de la suppression Changement prix double Affichage prix total Arrondi prix centième

+++ dead-pineapple-dal/src/com/deadpineapple/dal/entity/ConvertedFile.java
@@ -25,7 +25,7 @@ public class ConvertedFile implements Serializable {
     @JoinColumn(name = "userAccountId")
     private UserAccount userAccount;
 
-    private int size;
+    private double size;
 
     private Boolean isConverted;
 
@@ -59,11 +59,11 @@ public class ConvertedFile implements Serializable {
         this.splitFiles = splitFiles;
     }
 
-    public int getSize() {
+    public double getSize() {
         return size;
     }
 
-    public void setSize(int size) {
+    public void setSize(double size) {
         this.size = size;
     }
